# docs/swagger.yaml
definitions:
  main.Item:
    properties:
      description:
        type: string
      itemId:
        type: string
      quantity:
        type: integer
    type: object
  main.Order:
    properties:
      customerName:
        example: Bagus Aji
        type: string
      items:
        items:
          $ref: '#/definitions/main.Item'
        type: array
      orderId:
        type: string
      orderedAt:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: API Order
  version: "1.0"
paths:
  /orders:
    get:
      description: Get all orders description
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Order'
            type: array
      summary: Get all orders summary
      tags:
      - Order
    post:
      consumes:
      - application/json
      description: Get all orders description
      parameters:
      - description: Create Order
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/main.Order'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Order'
      summary: Get all orders summary
  /orders/{orderId}:
    get:
      description: Get all orders description
      parameters:
      - description: Get Order by Id
        in: path
        name: orderId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Order'
      summary: Get all orders summary
      tags:
      - Order
swagger: "2.0"